requirements.txt
Cython==0.27.3
Django==2.0.2
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==1.0.4
markdown2==2.3.5
pytz==2018.3
